django_app/requirements.txt
Django==2.1.7
django-filter==2.1.0
djangorestframework==3.9.2
gunicorn==19.9.0
Markdown==3.0.1
psycopg2-binary==2.7.7
pytz==2018.9
